--- src/compiler/types/lines_4000_8000.rs
@@ -1,1968 +1,0 @@
-#![allow(non_upper_case_globals)]
-
-use bitflags::bitflags;
-use std::cell::{Cell, Ref, RefCell, RefMut};
-use std::collections::HashMap;
-use std::convert::{TryFrom, TryInto};
-use std::fmt;
-use std::ops::BitAndAssign;
-use std::ops::Deref;
-use std::rc::{Rc, Weak};
-
-use super::{
-    DiagnosticCollection, ModuleSpecifierResolutionHost, Node, NodeId, PseudoBigInt,
-    RelationComparisonResult, SourceFile, SymbolTracker,
-};
-use crate::{NodeBuilder, Number, SortedArray, WeakSelf};
-use local_macros::{ast_type, enum_unwrapped, symbol_type, type_type};
-
-#[derive(Eq, PartialEq)]
-pub enum StructureIsReused {
-    Not,
-    Completely,
-}
-
-#[allow(non_camel_case_types)]
-pub enum ExitStatus {
-    Success,
-    DiagnosticsPresent_OutputsGenerated,
-}
-
-pub trait TypeCheckerHost: ModuleSpecifierResolutionHost {
-    fn get_source_files(&self) -> Vec<Rc<Node>>;
-}
-
-#[derive(Debug)]
-#[allow(non_snake_case)]
-pub struct TypeChecker {
-    pub _types_needing_strong_references: RefCell<Vec<Rc<Type>>>,
-    pub Symbol: fn(SymbolFlags, __String) -> BaseSymbol,
-    pub Type: fn(TypeFlags) -> BaseType,
-    pub(crate) type_count: Cell<u32>,
-    pub(crate) empty_symbols: Rc<RefCell<SymbolTable>>,
-    pub strict_null_checks: bool,
-    pub fresh_object_literal_flag: ObjectFlags,
-    pub exact_optional_property_types: bool,
-    pub node_builder: NodeBuilder,
-    pub globals: RefCell<SymbolTable>,
-    pub string_literal_types: RefCell<HashMap<String, Rc</*StringLiteralType*/ Type>>>,
-    pub number_literal_types: RefCell<HashMap<Number, Rc</*NumberLiteralType*/ Type>>>,
-    pub big_int_literal_types: RefCell<HashMap<String, Rc</*BigIntLiteralType*/ Type>>>,
-    pub unknown_symbol: Option<Rc<Symbol>>,
-    pub any_type: Option<Rc<Type>>,
-    pub error_type: Option<Rc<Type>>,
-    pub undefined_type: Option<Rc<Type>>,
-    pub null_type: Option<Rc<Type>>,
-    pub string_type: Option<Rc<Type>>,
-    pub number_type: Option<Rc<Type>>,
-    pub bigint_type: Option<Rc<Type>>,
-    pub true_type: Option<Rc<Type>>,
-    pub regular_true_type: Option<Rc<Type>>,
-    pub false_type: Option<Rc<Type>>,
-    pub regular_false_type: Option<Rc<Type>>,
-    pub boolean_type: Option<Rc<Type>>,
-    pub never_type: Option<Rc<Type>>,
-    pub number_or_big_int_type: Option<Rc<Type>>,
-    pub template_constraint_type: Option<Rc<Type>>,
-    pub global_array_type: Option<Rc<Type /*GenericType*/>>,
-    pub symbol_links: RefCell<HashMap<SymbolId, Rc<RefCell<SymbolLinks>>>>,
-    pub node_links: RefCell<HashMap<NodeId, Rc<RefCell<NodeLinks>>>>,
-    pub diagnostics: RefCell<DiagnosticCollection>,
-    pub assignable_relation: HashMap<String, RelationComparisonResult>,
-    pub comparable_relation: HashMap<String, RelationComparisonResult>,
-}
-
-bitflags! {
-    pub struct SymbolFormatFlags: u32 {
-        const None = 0;
-        const WriteTypeParametersOrArguments = 1 << 0;
-        const UseOnlyExternalAliasing = 1 << 1;
-        const AllowAnyNodeKind = 1 << 2;
-        const UseAliasDefinedOutsideCurrentScope = 1 << 3;
-        const DoNotIncludeSymbolChain = 1 << 4;
-    }
-}
-
-#[derive(PartialEq, Eq)]
-pub enum UnionReduction {
-    None,
-    Literal,
-    Subtype,
-}
-
-bitflags! {
-    pub struct NodeBuilderFlags: u32 {
-        const None = 0;
-        const NoTruncation = 1 << 0;
-        const WriteArrayAsGenericType = 1 << 1;
-        const UseStructuralFallback = 1 << 3;
-        const WriteTypeArgumentsOfSignature = 1 << 5;
-        const UseFullyQualifiedType = 1 << 6;
-        const UseOnlyExternalAliasing = 1 << 7;
-        const SuppressAnyReturnType = 1 << 8;
-        const WriteTypeParametersInQualifiedName = 1 << 9;
-        const MultilineObjectLiterals = 1 << 10;
-        const WriteClassExpressionAsTypeLiteral = 1 << 11;
-        const UseTypeOfFunction = 1 << 12;
-        const OmitParameterModifiers = 1 << 13;
-        const UseAliasDefinedOutsideCurrentScope = 1 << 14;
-        const UseSingleQuotesForStringLiteralType = 1 << 28;
-        const NoTypeReduction = 1 << 29;
-
-        const AllowThisInObjectLiteral = 1 << 15;
-        const AllowQualifiedNameInPlaceOfIdentifier = 1 << 16;
-        const AllowAnonymousIdentifier = 1 << 17;
-        const AllowEmptyUnionOrIntersection = 1 << 18;
-        const AllowEmptyTuple = 1 << 19;
-        const AllowUniqueESSymbolType = 1 << 20;
-        const AllowEmptyIndexInfoType = 1 << 21;
-
-        const AllowNodeModulesRelativePaths = 1 << 26;
-        const DoNotIncludeSymbolChain = 1 << 27;
-
-        const InTypeAlias = 1 << 23;
-
-        const IgnoreErrors = Self::AllowThisInObjectLiteral.bits | Self::AllowQualifiedNameInPlaceOfIdentifier.bits | Self::AllowAnonymousIdentifier.bits | Self::AllowEmptyUnionOrIntersection.bits | Self::AllowEmptyTuple.bits | Self::AllowEmptyIndexInfoType.bits | Self::AllowNodeModulesRelativePaths.bits;
-    }
-}
-
-bitflags! {
-    pub struct TypeFormatFlags: u32 {
-        const None = 0;
-        const NoTruncation = 1 << 0;
-        const WriteArrayAsGenericType = 1 << 1;
-        const UseStructuralFallback = 1 << 3;
-        const WriteTypeArgumentsOfSignature = 1 << 5;
-        const UseFullyQualifiedType = 1 << 6;
-        const SuppressAnyReturnType = 1 << 8;
-        const MultilineObjectLiterals = 1 << 10;
-        const WriteClassExpressionAsTypeLiteral = 1 << 11;
-        const UseTypeOfFunction = 1 << 12;
-        const OmitParameterModifiers = 1 << 13;
-
-        const UseAliasDefinedOutsideCurrentScope = 1 << 14;
-        const UseSingleQuotesForStringLiteralType = 1 << 28;
-        const NoTypeReduction = 1 << 29;
-
-        const AllowUniqueESSymbolType = 1 << 20;
-
-        const InTypeAlias = 1 << 23;
-
-        const NodeBuilderFlagsMask = Self::NoTruncation.bits | Self::WriteArrayAsGenericType.bits | Self::UseStructuralFallback.bits | Self::WriteTypeArgumentsOfSignature.bits | Self::UseFullyQualifiedType.bits | Self::SuppressAnyReturnType.bits | Self::MultilineObjectLiterals.bits | Self::WriteClassExpressionAsTypeLiteral.bits | Self::UseTypeOfFunction.bits | Self::OmitParameterModifiers.bits | Self::UseAliasDefinedOutsideCurrentScope.bits | Self::AllowUniqueESSymbolType.bits | Self::InTypeAlias.bits | Self::UseSingleQuotesForStringLiteralType.bits | Self::NoTypeReduction.bits;
-    }
-}
-
-pub trait SymbolWriter: SymbolTracker {
-    fn write_keyword(&mut self, text: &str);
-    fn write_punctuation(&mut self, text: &str);
-    fn write_space(&mut self, text: &str);
-    fn write_string_literal(&mut self, text: &str);
-    fn write_property(&mut self, text: &str);
-    fn write_symbol(&mut self, text: &str, symbol: &Symbol);
-    fn clear(&mut self);
-}
-
-bitflags! {
-    pub struct SymbolFlags: u32 {
-        const None = 0;
-        const FunctionScopedVariable = 1 << 0;
-        const BlockScopedVariable = 1 << 1;
-        const Property = 1 << 2;
-        const EnumMember = 1 << 3;
-        const Function = 1 << 4;
-        const Class = 1 << 5;
-        const Interface = 1 << 6;
-        const ConstEnum = 1 << 7;
-        const RegularEnum = 1 << 8;
-        const ValueModule = 1 << 9;
-        const TypeLiteral = 1 << 11;
-        const ObjectLiteral = 1 << 12;
-        const Method = 1 << 13;
-        const GetAccessor = 1 << 15;
-        const SetAccessor = 1 << 16;
-        const TypeParameter = 1 << 18;
-        const TypeAlias = 1 << 19;
-        const ExportValue = 1 << 20;
-        const Optional = 1 << 24;
-        const Transient = 1 << 25;
-
-        const Enum = Self::RegularEnum.bits | Self::ConstEnum.bits;
-        const Variable = Self::FunctionScopedVariable.bits | Self::BlockScopedVariable.bits;
-        const Value = Self::Variable.bits | Self::Property.bits | Self::EnumMember.bits | Self::ObjectLiteral.bits | Self::Function.bits | Self::Class.bits | Self::Enum.bits | Self::ValueModule.bits | Self::Method.bits | Self::GetAccessor.bits | Self::SetAccessor.bits;
-        const Type = Self::Class.bits | Self::Interface.bits | Self::Enum.bits | Self::EnumMember.bits | Self::TypeLiteral.bits | Self::TypeParameter.bits | Self::TypeAlias.bits;
-
-        const FunctionScopedVariableExcludes = Self::Value.bits & !Self::FunctionScopedVariable.bits;
-
-        const BlockScopedVariableExcludes = Self::Value.bits;
-
-        const ParameterExcludes = Self::Value.bits;
-        const PropertyExcludes = Self::None.bits;
-        const FunctionExcludes = Self::Value.bits & !(Self::Function.bits | Self::ValueModule.bits | Self::Class.bits);
-        const InterfaceExcludes = Self::Type.bits & !(Self::Interface.bits | Self::Class.bits);
-        const TypeParameterExcludes = Self::Type.bits & !Self::TypeParameter.bits;
-        const TypeAliasExcludes = Self::Type.bits;
-    }
-}
-
-pub trait SymbolInterface {
-    fn symbol_wrapper(&self) -> Rc<Symbol>;
-    fn set_symbol_wrapper(&self, wrapper: Rc<Symbol>);
-    fn flags(&self) -> SymbolFlags;
-    fn set_flags(&self, flags: SymbolFlags);
-    fn escaped_name(&self) -> &__String;
-    fn maybe_declarations(&self) -> Ref<Option<Vec<Rc<Node>>>>;
-    fn set_declarations(&self, declarations: Vec<Rc<Node>>);
-    fn maybe_value_declaration(&self) -> Ref<Option<Weak<Node>>>;
-    fn set_value_declaration(&self, node: Rc<Node>);
-    fn maybe_members(&self) -> RefMut<Option<Rc<RefCell<SymbolTable>>>>;
-    fn members(&self) -> Rc<RefCell<SymbolTable>>;
-    fn maybe_id(&self) -> Option<SymbolId>;
-    fn id(&self) -> SymbolId;
-    fn set_id(&self, id: SymbolId);
-}
-
-pub type SymbolId = u32;
-
-#[derive(Debug)]
-#[symbol_type(impl_from = false)]
-pub enum Symbol {
-    BaseSymbol(BaseSymbol),
-    TransientSymbol(TransientSymbol),
-}
-
-impl Symbol {
-    pub fn wrap(self) -> Rc<Symbol> {
-        let rc = Rc::new(self);
-        rc.set_symbol_wrapper(rc.clone());
-        rc
-    }
-}
-
-#[derive(Debug)]
-pub struct BaseSymbol {
-    _symbol_wrapper: RefCell<Option<Weak<Symbol>>>,
-    flags: Cell<SymbolFlags>,
-    escaped_name: __String,
-    declarations: RefCell<Option<Vec<Rc<Node /*Declaration*/>>>>, // TODO: should be Vec<Weak<Node>> instead of Vec<Rc<Node>>?
-    value_declaration: RefCell<Option<Weak<Node>>>,
-    members: RefCell<Option<Rc<RefCell<SymbolTable>>>>,
-    id: Cell<Option<SymbolId>>,
-}
-
-impl BaseSymbol {
-    pub fn new(flags: SymbolFlags, name: __String) -> Self {
-        Self {
-            _symbol_wrapper: RefCell::new(None),
-            flags: Cell::new(flags),
-            escaped_name: name,
-            declarations: RefCell::new(None),
-            value_declaration: RefCell::new(None),
-            members: RefCell::new(None),
-            id: Cell::new(None),
-        }
-    }
-}
-
-impl SymbolInterface for BaseSymbol {
-    fn symbol_wrapper(&self) -> Rc<Symbol> {
-        self._symbol_wrapper
-            .borrow()
-            .as_ref()
-            .unwrap()
-            .upgrade()
-            .unwrap()
-    }
-
-    fn set_symbol_wrapper(&self, wrapper: Rc<Symbol>) {
-        *self._symbol_wrapper.borrow_mut() = Some(Rc::downgrade(&wrapper));
-    }
-
-    fn flags(&self) -> SymbolFlags {
-        self.flags.get()
-    }
-
-    fn set_flags(&self, flags: SymbolFlags) {
-        self.flags.set(flags);
-    }
-
-    fn escaped_name(&self) -> &__String {
-        &self.escaped_name
-    }
-
-    fn maybe_declarations(&self) -> Ref<Option<Vec<Rc<Node>>>> {
-        self.declarations.borrow()
-    }
-
-    fn set_declarations(&self, declarations: Vec<Rc<Node>>) {
-        *self.declarations.borrow_mut() = Some(declarations);
-    }
-
-    fn maybe_value_declaration(&self) -> Ref<Option<Weak<Node>>> {
-        self.value_declaration.borrow()
-    }
-
-    fn set_value_declaration(&self, node: Rc<Node>) {
-        *self.value_declaration.borrow_mut() = Some(Rc::downgrade(&node));
-    }
-
-    fn maybe_members(&self) -> RefMut<Option<Rc<RefCell<SymbolTable>>>> {
-        self.members.borrow_mut()
-    }
-
-    fn members(&self) -> Rc<RefCell<SymbolTable>> {
-        self.members.borrow_mut().as_ref().unwrap().clone()
-    }
-
-    fn maybe_id(&self) -> Option<SymbolId> {
-        self.id.get()
-    }
-
-    fn id(&self) -> SymbolId {
-        self.id.get().unwrap()
-    }
-
-    fn set_id(&self, id: SymbolId) {
-        self.id.set(Some(id));
-    }
-}
-
-impl From<BaseSymbol> for Symbol {
-    fn from(base_symbol: BaseSymbol) -> Self {
-        Symbol::BaseSymbol(base_symbol)
-    }
-}
-
-#[derive(Debug)]
-pub struct SymbolLinks {
-    pub target: Option<Rc<Symbol>>,
-    pub type_: Option<Rc<Type>>,
-    pub declared_type: Option<Rc<Type>>,
-    pub mapper: Option<TypeMapper>,
-}
-
-impl SymbolLinks {
-    pub fn new() -> Self {
-        Self {
-            target: None,
-            type_: None,
-            declared_type: None,
-            mapper: None,
-        }
-    }
-}
-
-bitflags! {
-    pub struct CheckFlags: u32 {
-        const None = 0;
-        const Instantiated = 1 << 0;
-        const SyntheticProperty = 1 << 1;
-        const SyntheticMethod = 1 << 2;
-        const Readonly = 1 << 3;
-        const Late = 1 << 12;
-        const OptionalParameter = 1 << 14;
-        const RestParameter = 1 << 15;
-
-        const Synthetic = Self::SyntheticProperty.bits | Self::SyntheticMethod.bits;
-    }
-}
-
-pub trait TransientSymbolInterface: SymbolInterface {
-    fn symbol_links(&self) -> Rc<RefCell<SymbolLinks>>;
-    fn check_flags(&self) -> CheckFlags;
-}
-
-#[derive(Debug)]
-#[symbol_type(interfaces = "TransientSymbolInterface")]
-pub enum TransientSymbol {
-    BaseTransientSymbol(BaseTransientSymbol),
-}
-
-#[derive(Debug)]
-#[symbol_type(ancestors = "TransientSymbol")]
-pub struct BaseTransientSymbol {
-    _symbol: BaseSymbol,
-    _symbol_links: Rc<RefCell<SymbolLinks>>,
-    check_flags: CheckFlags,
-}
-
-impl BaseTransientSymbol {
-    pub fn new(base_symbol: BaseSymbol, check_flags: CheckFlags) -> Self {
-        Self {
-            _symbol: base_symbol,
-            _symbol_links: Rc::new(RefCell::new(SymbolLinks::new())),
-            check_flags,
-        }
-    }
-}
-
-impl TransientSymbolInterface for BaseTransientSymbol {
-    fn symbol_links(&self) -> Rc<RefCell<SymbolLinks>> {
-        self._symbol_links.clone()
-    }
-
-    fn check_flags(&self) -> CheckFlags {
-        self.check_flags
-    }
-}
-
-pub struct InternalSymbolName;
-
-#[allow(non_snake_case)]
-impl InternalSymbolName {
-    pub fn Object() -> __String {
-        __String::new("__object".to_string())
-    }
-}
-
-#[derive(Clone, Debug, Hash, Eq, PartialEq)]
-pub struct __String(String);
-
-impl __String {
-    pub fn new(string: String) -> Self {
-        Self(string)
-    }
-
-    pub fn chars(&self) -> std::str::Chars {
-        self.0.chars()
-    }
-}
-
-pub type UnderscoreEscapedMap<TValue> = HashMap<__String, TValue>;
-
-pub type SymbolTable = UnderscoreEscapedMap<Rc<Symbol>>;
-
-bitflags! {
-    pub struct NodeCheckFlags: u32 {
-        const None = 0;
-    }
-}
-
-#[derive(Debug)]
-pub struct NodeLinks {
-    pub flags: NodeCheckFlags,
-    pub resolved_type: Option<Rc<Type>>,
-    pub resolved_symbol: Option<Rc<Symbol>>,
-}
-
-impl NodeLinks {
-    pub fn new() -> Self {
-        Self {
-            flags: NodeCheckFlags::None,
-            resolved_type: None,
-            resolved_symbol: None,
-        }
-    }
-}
-
-bitflags! {
-    pub struct TypeFlags: u32 {
-        const None = 0;
-        const Any = 1 << 0;
-        const Unknown = 1 << 1;
-        const String = 1 << 2;
-        const Number = 1 << 3;
-        const Boolean = 1 << 4;
-        const Enum = 1 << 5;
-        const BigInt = 1 << 6;
-        const StringLiteral = 1 << 7;
-        const NumberLiteral = 1 << 8;
-        const BooleanLiteral = 1 << 9;
-        const EnumLiteral = 1 << 10;
-        const BigIntLiteral = 1 << 11;
-        const ESSymbol = 1 << 12;
-        const UniqueESSymbol = 1 << 13;
-        const Void = 1 << 14;
-        const Undefined = 1 << 15;
-        const Null = 1 << 16;
-        const Never = 1 << 17;
-        const TypeParameter = 1 << 18;
-        const Object = 1 << 19;
-        const Union = 1 << 20;
-        const Intersection = 1 << 21;
-        const Index = 1 << 22;
-        const IndexedAccess = 1 << 23;
-        const Conditional = 1 << 24;
-        const Substitution = 1 << 25;
-        const NonPrimitive = 1 << 26;
-        const TemplateLiteral = 1 << 27;
-        const StringMapping = 1 << 28;
-
-        const Nullable = Self::Undefined.bits | Self::Null.bits;
-        const Literal = Self::StringLiteral.bits | Self::NumberLiteral.bits | Self::BigIntLiteral.bits | Self::BooleanLiteral.bits;
-        const Unit = Self::Literal.bits | Self::UniqueESSymbol.bits | Self::Nullable.bits;
-        const StringOrNumberLiteralOrUnique = Self::StringLiteral.bits | Self::NumberLiteral.bits | Self::UniqueESSymbol.bits;
-        const Primitive = Self::String.bits | Self::Number.bits | Self::BigInt.bits | Self::Boolean.bits | Self::Enum.bits | Self::EnumLiteral.bits | Self::ESSymbol.bits | Self::Void.bits | Self::Undefined.bits | Self::Null.bits | Self::Literal.bits | Self::UniqueESSymbol.bits;
-        const NumberLike = Self::Number.bits | Self::NumberLiteral.bits | Self::Enum.bits;
-        const BigIntLike = Self::BigInt.bits | Self::BigIntLiteral.bits;
-        const UnionOrIntersection =  Self::Union.bits | Self::Intersection.bits;
-        const StructuredType = Self::Object.bits | Self::Union.bits | Self::Intersection.bits;
-        const TypeVariable = Self::TypeParameter.bits | Self::IndexedAccess.bits;
-        const InstantiableNonPrimitive = Self::TypeVariable.bits | Self::Conditional.bits | Self::Substitution.bits;
-        const InstantiablePrimitive = Self::Index.bits | Self::TemplateLiteral.bits | Self::StringMapping.bits;
-        const Instantiable = Self::InstantiableNonPrimitive.bits | Self::InstantiablePrimitive.bits;
-        const StructuredOrInstantiable = Self::StructuredType.bits | Self::Instantiable.bits;
-        const ObjectFlagsType = Self::Any.bits | Self::Nullable.bits | Self::Never.bits | Self::Object.bits | Self::Union.bits | Self::Intersection.bits;
-        const IncludesMask = Self::Any.bits | Self::Unknown.bits | Self::Primitive.bits | Self::Never.bits | Self::Object.bits | Self::Union.bits | Self::Intersection.bits | Self::NonPrimitive.bits | Self::TemplateLiteral.bits;
-        const IncludesInstantiable = Self::Substitution.bits;
-        const NotPrimitiveUnion = Self::Any.bits | Self::Unknown.bits | Self::Enum.bits | Self::Void.bits | Self::Never.bits | Self::Object.bits | Self::Intersection.bits | Self::IncludesInstantiable.bits;
-    }
-}
-
-pub type TypeId = u32;
-
-#[derive(Clone, Debug)]
-#[type_type(impl_from = false)]
-pub enum Type {
-    IntrinsicType(IntrinsicType),
-    LiteralType(LiteralType),
-    ObjectType(ObjectType),
-    UnionOrIntersectionType(UnionOrIntersectionType),
-    TypeParameter(TypeParameter),
-}
-
-impl Type {
-    pub fn as_intrinsic_type(&self) -> &dyn IntrinsicTypeInterface {
-        match self {
-            Type::IntrinsicType(intrinsic_type) => intrinsic_type,
-            _ => panic!("Expected intrinsic type"),
-        }
-    }
-
-    pub fn as_union_or_intersection_type_interface(&self) -> &dyn UnionOrIntersectionTypeInterface {
-        match self {
-            Type::UnionOrIntersectionType(union_or_intersection_type) => union_or_intersection_type,
-            _ => panic!("Expected union or intersection type"),
-        }
-    }
-
-    pub fn as_resolved_type(&self) -> &dyn ResolvedTypeInterface {
-        match self {
-            Type::ObjectType(object_type) => {
-                if !object_type.is_resolved() {
-                    panic!("Not resolved")
-                }
-                object_type
-            }
-            _ => panic!("Expected resolved type"),
-        }
-    }
-
-    pub fn as_resolvable_type(&self) -> &dyn ResolvableTypeInterface {
-        match self {
-            Type::ObjectType(object_type) => object_type,
-            _ => panic!("Expected resolvable type"),
-        }
-    }
-
-    pub fn as_object_flags_type(&self) -> &dyn ObjectFlagsTypeInterface {
-        match self {
-            Type::ObjectType(object_type) => object_type,
-            Type::UnionOrIntersectionType(union_or_intersection_type) => union_or_intersection_type,
-            _ => panic!("Expected object flags type"),
-        }
-    }
-
-    pub fn as_base_interface_type(&self) -> &BaseInterfaceType {
-        enum_unwrapped!(self, [Type, ObjectType, InterfaceType, BaseInterfaceType])
-    }
-
-    pub fn as_type_parameter(&self) -> &TypeParameter {
-        enum_unwrapped!(self, [Type, TypeParameter])
-    }
-
-    pub fn as_interface_type(&self) -> &InterfaceType {
-        enum_unwrapped!(self, [Type, ObjectType, InterfaceType])
-    }
-
-    pub fn as_object_type(&self) -> &ObjectType {
-        enum_unwrapped!(self, [Type, ObjectType])
-    }
-
-    pub fn as_type_reference(&self) -> &TypeReference {
-        enum_unwrapped!(self, [Type, ObjectType, TypeReference])
-    }
-
-    pub fn as_string_literal_type(&self) -> &StringLiteralType {
-        enum_unwrapped!(self, [Type, LiteralType, StringLiteralType])
-    }
-
-    pub fn as_number_literal_type(&self) -> &NumberLiteralType {
-        enum_unwrapped!(self, [Type, LiteralType, NumberLiteralType])
-    }
-
-    pub fn as_big_int_literal_type(&self) -> &BigIntLiteralType {
-        enum_unwrapped!(self, [Type, LiteralType, BigIntLiteralType])
-    }
-
-    pub fn as_union_or_intersection_type(&self) -> &UnionOrIntersectionType {
-        enum_unwrapped!(self, [Type, UnionOrIntersectionType])
-    }
-
-    pub fn as_literal_type(&self) -> &LiteralType {
-        enum_unwrapped!(self, [Type, LiteralType])
-    }
-
-    pub fn as_freshable_intrinsic_type(&self) -> &FreshableIntrinsicType {
-        enum_unwrapped!(self, [Type, IntrinsicType, FreshableIntrinsicType])
-    }
-}
-
-pub trait TypeInterface {
-    fn type_wrapper(&self) -> Rc<Type>;
-    fn set_type_wrapper(&self, wrapper: Rc<Type>);
-    fn flags(&self) -> TypeFlags;
-    fn id(&self) -> TypeId;
-    fn maybe_symbol(&self) -> Option<Rc<Symbol>>;
-    fn symbol(&self) -> Rc<Symbol>;
-    fn set_symbol(&mut self, symbol: Rc<Symbol>);
-}
-
-#[derive(Clone, Debug)]
-pub struct BaseType {
-    _type_wrapper: RefCell<Option<Weak<Type>>>,
-    pub flags: TypeFlags,
-    pub id: Option<TypeId>,
-    symbol: Option<Rc<Symbol>>,
-}
-
-impl BaseType {
-    pub fn new(flags: TypeFlags) -> Self {
-        Self {
-            _type_wrapper: RefCell::new(None),
-            flags,
-            id: None,
-            symbol: None,
-        }
-    }
-}
-
-impl TypeInterface for BaseType {
-    fn type_wrapper(&self) -> Rc<Type> {
-        self._type_wrapper
-            .borrow()
-            .as_ref()
-            .unwrap()
-            .upgrade()
-            .unwrap()
-    }
-
-    fn set_type_wrapper(&self, wrapper: Rc<Type>) {
-        *self._type_wrapper.borrow_mut() = Some(Rc::downgrade(&wrapper));
-    }
-
-    fn flags(&self) -> TypeFlags {
-        self.flags
-    }
-
-    fn id(&self) -> TypeId {
-        self.id.unwrap()
-    }
-
-    fn maybe_symbol(&self) -> Option<Rc<Symbol>> {
-        self.symbol.as_ref().map(Clone::clone)
-    }
-
-    fn symbol(&self) -> Rc<Symbol> {
-        self.symbol.as_ref().unwrap().clone()
-    }
-
-    fn set_symbol(&mut self, symbol: Rc<Symbol>) {
-        self.symbol = Some(symbol);
-    }
-}
-
-pub trait IntrinsicTypeInterface: TypeInterface {
-    fn intrinsic_name(&self) -> &str;
-}
-
-#[derive(Clone, Debug)]
-#[type_type(interfaces = "IntrinsicTypeInterface")]
-pub enum IntrinsicType {
-    BaseIntrinsicType(BaseIntrinsicType),
-    FreshableIntrinsicType(FreshableIntrinsicType),
-}
-
-#[derive(Clone, Debug)]
-#[type_type(ancestors = "IntrinsicType")]
-pub struct BaseIntrinsicType {
-    _type: BaseType,
-    intrinsic_name: String,
-}
-
-impl BaseIntrinsicType {
-    pub fn new(type_: BaseType, intrinsic_name: String) -> Self {
-        Self {
-            _type: type_,
-            intrinsic_name,
-        }
-    }
-}
-
-impl IntrinsicTypeInterface for BaseIntrinsicType {
-    fn intrinsic_name(&self) -> &str {
-        &self.intrinsic_name
-    }
-}
-
-#[derive(Clone, Debug)]
-#[type_type(ancestors = "IntrinsicType", interfaces = "IntrinsicTypeInterface")]
-pub struct FreshableIntrinsicType {
-    _intrinsic_type: BaseIntrinsicType,
-    pub fresh_type: WeakSelf<Type>,
-    pub regular_type: WeakSelf<Type>,
-}
-
-impl FreshableIntrinsicType {
-    pub fn new(intrinsic_type: BaseIntrinsicType) -> Self {
-        Self {
-            _intrinsic_type: intrinsic_type,
-            fresh_type: WeakSelf::new(),
-            regular_type: WeakSelf::new(),
-        }
-    }
-
-    pub fn fresh_type(&self) -> Weak<Type> {
-        self.fresh_type.get()
-    }
-
-    pub fn regular_type(&self) -> Weak<Type> {
-        self.regular_type.get()
-    }
-}
-
-pub trait LiteralTypeInterface: TypeInterface {
-    fn fresh_type(&self) -> Option<&Weak<Type>>;
-    fn set_fresh_type(&self, fresh_type: &Rc<Type>);
-    fn get_or_initialize_fresh_type(&self, type_checker: &TypeChecker) -> Rc<Type>;
-    fn regular_type(&self) -> Rc<Type>;
-    fn set_regular_type(&self, regular_type: &Rc<Type>);
-}
-
-#[derive(Clone, Debug)]
-#[type_type(interfaces = "LiteralTypeInterface")]
-pub enum LiteralType {
-    StringLiteralType(StringLiteralType),
-    NumberLiteralType(NumberLiteralType),
-    BigIntLiteralType(BigIntLiteralType),
-}
-
-#[derive(Clone, Debug)]
-#[type_type(impl_from = false)]
-pub struct BaseLiteralType {
-    _type: BaseType,
-    fresh_type: WeakSelf<Type>,
-    regular_type: WeakSelf<Type>,
-}
-
-impl BaseLiteralType {
-    pub fn new(type_: BaseType) -> Self {
-        Self {
-            _type: type_,
-            fresh_type: WeakSelf::new(),
-            regular_type: WeakSelf::new(),
-        }
-    }
-}
-
-impl LiteralTypeInterface for BaseLiteralType {
-    fn fresh_type(&self) -> Option<&Weak<Type>> {
-        self.fresh_type.try_get()
-    }
-
-    fn set_fresh_type(&self, fresh_type: &Rc<Type>) {
-        self.fresh_type.init(fresh_type, false);
-    }
-
-    fn get_or_initialize_fresh_type(&self, type_checker: &TypeChecker) -> Rc<Type> {
-        panic!("Shouldn't call get_or_initialize_fresh_type() on base BaseLiteralType");
-    }
-
-    fn regular_type(&self) -> Rc<Type> {
-        self.regular_type.get().upgrade().unwrap()
-    }
-
-    fn set_regular_type(&self, regular_type: &Rc<Type>) {
-        self.regular_type.init(regular_type, false);
-    }
-}
-
-#[derive(Clone, Debug)]
-#[type_type(ancestors = "LiteralType")]
-pub struct StringLiteralType {
-    _literal_type: BaseLiteralType,
-    pub value: String,
-}
-
-impl StringLiteralType {
-    pub fn new(literal_type: BaseLiteralType, value: String) -> Self {
-        Self {
-            _literal_type: literal_type,
-            value,
-        }
-    }
-
-    fn create_fresh_type_from_self(&self, type_checker: &TypeChecker) -> Rc<Type> {
-        let fresh_type = type_checker.create_string_literal_type(
-            self.flags(),
-            self.value.clone(),
-            Some(self.type_wrapper()),
-        );
-        fresh_type.as_literal_type().set_fresh_type(&fresh_type);
-        self.set_fresh_type(&fresh_type);
-        type_checker.keep_strong_reference_to_type(fresh_type);
-        self.fresh_type().unwrap().upgrade().unwrap()
-    }
-}
-
-impl LiteralTypeInterface for StringLiteralType {
-    fn fresh_type(&self) -> Option<&Weak<Type>> {
-        self._literal_type.fresh_type()
-    }
-
-    fn set_fresh_type(&self, fresh_type: &Rc<Type>) {
-        self._literal_type.set_fresh_type(fresh_type);
-    }
-
-    fn get_or_initialize_fresh_type(&self, type_checker: &TypeChecker) -> Rc<Type> {
-        if self.fresh_type().is_none() {
-            let fresh_type = self.create_fresh_type_from_self(type_checker);
-            self.set_fresh_type(&fresh_type);
-            return self.fresh_type().unwrap().upgrade().unwrap();
-        }
-        return self.fresh_type().unwrap().upgrade().unwrap();
-    }
-
-    fn regular_type(&self) -> Rc<Type> {
-        self._literal_type.regular_type()
-    }
-
-    fn set_regular_type(&self, regular_type: &Rc<Type>) {
-        self._literal_type.set_regular_type(regular_type);
-    }
-}
-
-#[derive(Clone, Debug)]
-#[type_type(ancestors = "LiteralType")]
-pub struct NumberLiteralType {
-    _literal_type: BaseLiteralType,
-    pub value: Number,
-}
-
-impl NumberLiteralType {
-    pub fn new(literal_type: BaseLiteralType, value: Number) -> Self {
-        Self {
-            _literal_type: literal_type,
-            value,
-        }
-    }
-
-    fn create_fresh_type_from_self(&self, type_checker: &TypeChecker) -> Rc<Type> {
-        let fresh_type = type_checker.create_number_literal_type(
-            self.flags(),
-            self.value,
-            Some(self.type_wrapper()),
-        );
-        fresh_type.as_literal_type().set_fresh_type(&fresh_type);
-        self.set_fresh_type(&fresh_type);
-        type_checker.keep_strong_reference_to_type(fresh_type);
-        self.fresh_type().unwrap().upgrade().unwrap()
-    }
-}
-
-impl LiteralTypeInterface for NumberLiteralType {
-    fn fresh_type(&self) -> Option<&Weak<Type>> {
-        self._literal_type.fresh_type()
-    }
-
-    fn set_fresh_type(&self, fresh_type: &Rc<Type>) {
-        self._literal_type.set_fresh_type(fresh_type);
-    }
-
-    fn get_or_initialize_fresh_type(&self, type_checker: &TypeChecker) -> Rc<Type> {
-        if self.fresh_type().is_none() {
-            let fresh_type = self.create_fresh_type_from_self(type_checker);
-            self.set_fresh_type(&fresh_type);
-            return self.fresh_type().unwrap().upgrade().unwrap();
-        }
-        return self.fresh_type().unwrap().upgrade().unwrap();
-    }
-
-    fn regular_type(&self) -> Rc<Type> {
-        self._literal_type.regular_type()
-    }
-
-    fn set_regular_type(&self, regular_type: &Rc<Type>) {
-        self._literal_type.set_regular_type(regular_type);
-    }
-}
-
-#[derive(Clone, Debug)]
-#[type_type(ancestors = "LiteralType")]
-pub struct BigIntLiteralType {
-    _literal_type: BaseLiteralType,
-    pub value: PseudoBigInt,
-}
-
-impl BigIntLiteralType {
-    pub fn new(literal_type: BaseLiteralType, value: PseudoBigInt) -> Self {
-        Self {
-            _literal_type: literal_type,
-            value,
-        }
-    }
-
-    fn create_fresh_type_from_self(&self, type_checker: &TypeChecker) -> Rc<Type> {
-        let fresh_type = type_checker.create_big_int_literal_type(
-            self.flags(),
-            self.value.clone(),
-            Some(self.type_wrapper()),
-        );
-        fresh_type.as_literal_type().set_fresh_type(&fresh_type);
-        self.set_fresh_type(&fresh_type);
-        type_checker.keep_strong_reference_to_type(fresh_type);
-        self.fresh_type().unwrap().upgrade().unwrap()
-    }
-}
-
-impl LiteralTypeInterface for BigIntLiteralType {
-    fn fresh_type(&self) -> Option<&Weak<Type>> {
-        self._literal_type.fresh_type()
-    }
-
-    fn set_fresh_type(&self, fresh_type: &Rc<Type>) {
-        self._literal_type.set_fresh_type(fresh_type);
-    }
-
-    fn get_or_initialize_fresh_type(&self, type_checker: &TypeChecker) -> Rc<Type> {
-        if self.fresh_type().is_none() {
-            let fresh_type = self.create_fresh_type_from_self(type_checker);
-            self.set_fresh_type(&fresh_type);
-            return self.fresh_type().unwrap().upgrade().unwrap();
-        }
-        return self.fresh_type().unwrap().upgrade().unwrap();
-    }
-
-    fn regular_type(&self) -> Rc<Type> {
-        self._literal_type.regular_type()
-    }
-
-    fn set_regular_type(&self, regular_type: &Rc<Type>) {
-        self._literal_type.set_regular_type(regular_type);
-    }
-}
-
-bitflags! {
-    pub struct ObjectFlags: u32 {
-        const None = 0;
-        const Class = 1 << 0;
-        const Interface = 1 << 1;
-        const Reference = 1 << 2;
-        const Anonymous = 1 << 4;
-        const Mapped = 1 << 5;
-        const ObjectLiteral = 1 << 7;
-        const ObjectLiteralPatternWithComputedProperties = 1 << 9;
-        const FreshLiteral = 1 << 14;
-        const PrimitiveUnion = 1 << 16;
-        const ContainsWideningType = 1 << 17;
-        const ContainsObjectOrArrayLiteral = 1 << 18;
-        const NonInferrableType = 1 << 19;
-        const CouldContainTypeVariablesComputed = 1 << 20;
-        const CouldContainTypeVariables = 1 << 21;
-        const ContainsIntersections = 1 << 25;
-
-        const ClassOrInterface = Self::Class.bits | Self::Interface.bits;
-        const PropagatingFlags = Self::ContainsWideningType.bits | Self::ContainsObjectOrArrayLiteral.bits | Self::NonInferrableType.bits;
-    }
-}
-
-pub trait ObjectFlagsTypeInterface {
-    fn object_flags(&self) -> ObjectFlags;
-    fn set_object_flags(&self, object_flags: ObjectFlags);
-}
-
-pub trait ObjectTypeInterface: ObjectFlagsTypeInterface {
-    // fn maybe_properties(&self) -> Option<&[Rc<Symbol>]>;
-    // fn properties(&self) -> &[Rc<Symbol>];
-    // fn set_properties(&self, properties: Vec<Rc<Symbol>>);
-}
-
-#[derive(Clone, Debug)]
-#[type_type(
-    interfaces = "ObjectFlagsTypeInterface, ObjectTypeInterface, ResolvableTypeInterface, ResolvedTypeInterface"
-)]
-pub enum ObjectType {
-    BaseObjectType(BaseObjectType),
-    InterfaceType(InterfaceType),
-    TypeReference(TypeReference),
-}
-
-#[derive(Clone, Debug)]
-#[type_type(ancestors = "ObjectType")]
-pub struct BaseObjectType {
-    _type: BaseType,
-    object_flags: Cell<ObjectFlags>,
-    members: RefCell<Option<Rc<RefCell<SymbolTable>>>>,
-    properties: RefCell<Option<Vec<Rc<Symbol>>>>,
-}
-
-impl BaseObjectType {
-    pub fn new(base_type: BaseType, object_flags: ObjectFlags) -> Self {
-        Self {
-            _type: base_type,
-            object_flags: Cell::new(object_flags),
-            members: RefCell::new(None),
-            properties: RefCell::new(None),
-        }
-    }
-}
-
-impl ObjectFlagsTypeInterface for BaseObjectType {
-    fn object_flags(&self) -> ObjectFlags {
-        self.object_flags.get()
-    }
-
-    fn set_object_flags(&self, object_flags: ObjectFlags) {
-        self.object_flags.set(object_flags);
-    }
-}
-
-impl ObjectTypeInterface for BaseObjectType {}
-
-pub trait ResolvableTypeInterface {
-    fn resolve(&self, members: Rc<RefCell<SymbolTable>>, properties: Vec<Rc<Symbol>>);
-    fn is_resolved(&self) -> bool;
-}
-
-impl ResolvableTypeInterface for BaseObjectType {
-    fn resolve(&self, members: Rc<RefCell<SymbolTable>>, properties: Vec<Rc<Symbol>>) {
-        *self.members.borrow_mut() = Some(members);
-        *self.properties.borrow_mut() = Some(properties);
-    }
-
-    fn is_resolved(&self) -> bool {
-        self.members.borrow().is_some()
-    }
-}
-
-impl ResolvedTypeInterface for BaseObjectType {
-    fn members(&self) -> Rc<RefCell<SymbolTable>> {
-        self.members.borrow_mut().as_ref().unwrap().clone()
-    }
-
-    fn properties(&self) -> RefMut<Vec<Rc<Symbol>>> {
-        RefMut::map(self.properties.borrow_mut(), |option| {
-            option.as_mut().unwrap()
-        })
-    }
-
-    fn set_properties(&self, properties: Vec<Rc<Symbol>>) {
-        *self.properties.borrow_mut() = Some(properties);
-    }
-}
-
-#[derive(Clone, Debug)]
-#[type_type(
-    ancestors = "ObjectType",
-    interfaces = "ObjectFlagsTypeInterface, ObjectTypeInterface, ResolvableTypeInterface, ResolvedTypeInterface"
-)]
-pub enum InterfaceType {
-    BaseInterfaceType(BaseInterfaceType),
-}
-
-impl InterfaceTypeWithDeclaredMembersInterface for InterfaceType {
-    fn maybe_declared_properties(&self) -> Ref<Option<Vec<Rc<Symbol>>>> {
-        match self {
-            InterfaceType::BaseInterfaceType(base_interface_type) => {
-                base_interface_type.maybe_declared_properties()
-            }
-        }
-    }
-
-    fn set_declared_properties(&self, declared_properties: Vec<Rc<Symbol>>) {
-        match self {
-            InterfaceType::BaseInterfaceType(base_interface_type) => {
-                base_interface_type.set_declared_properties(declared_properties)
-            }
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-#[type_type(
-    ancestors = "InterfaceType, ObjectType",
-    interfaces = "ObjectFlagsTypeInterface, ObjectTypeInterface, ResolvableTypeInterface, ResolvedTypeInterface"
-)]
-pub struct BaseInterfaceType {
-    _object_type: BaseObjectType,
-    pub type_parameters: Option<Vec<Rc<Type /*TypeParameter*/>>>,
-    pub outer_type_parameters: Option<Vec<Rc<Type /*TypeParameter*/>>>,
-    pub local_type_parameters: Option<Vec<Rc<Type /*TypeParameter*/>>>,
-    pub this_type: RefCell<Option<Rc<Type /*TypeParameter*/>>>,
-    declared_properties: RefCell<Option<Vec<Rc<Symbol>>>>,
-}
-
-impl BaseInterfaceType {
-    pub fn new(
-        object_type: BaseObjectType,
-        type_parameters: Option<Vec<Rc<Type>>>,
-        outer_type_parameters: Option<Vec<Rc<Type>>>,
-        local_type_parameters: Option<Vec<Rc<Type>>>,
-        this_type: Option<Rc<Type>>,
-    ) -> Self {
-        Self {
-            _object_type: object_type,
-            type_parameters,
-            outer_type_parameters,
-            local_type_parameters,
-            this_type: RefCell::new(this_type),
-            declared_properties: RefCell::new(None),
-        }
-    }
-}
-
-impl InterfaceTypeWithDeclaredMembersInterface for BaseInterfaceType {
-    fn maybe_declared_properties(&self) -> Ref<Option<Vec<Rc<Symbol>>>> {
-        self.declared_properties.borrow()
-    }
-
-    fn set_declared_properties(&self, declared_properties: Vec<Rc<Symbol>>) {
-        *self.declared_properties.borrow_mut() = Some(declared_properties);
-    }
-}
-
-pub trait InterfaceTypeWithDeclaredMembersInterface {
-    fn maybe_declared_properties(&self) -> Ref<Option<Vec<Rc<Symbol>>>>;
-    fn set_declared_properties(&self, declared_properties: Vec<Rc<Symbol>>);
-}
-
-#[derive(Clone, Debug)]
-#[type_type(
-    ancestors = "ObjectType",
-    interfaces = "ObjectFlagsTypeInterface, ObjectTypeInterface, ResolvableTypeInterface, ResolvedTypeInterface"
-)]
-pub struct TypeReference {
-    _object_type: BaseObjectType,
-    pub target: Rc<Type /*GenericType*/>,
-    pub node: RefCell<Option<Rc<Node /*TypeReferenceNode | ArrayTypeNode | TupleTypeNode*/>>>, // TODO: should be weak?
-    pub resolved_type_arguments: RefCell<Option<Vec<Rc<Type>>>>,
-}
-
-impl TypeReference {
-    pub fn new(
-        object_type: BaseObjectType,
-        target: Rc<Type>,
-        resolved_type_arguments: Option<Vec<Rc<Type>>>,
-    ) -> Self {
-        Self {
-            _object_type: object_type,
-            target,
-            node: RefCell::new(None),
-            resolved_type_arguments: RefCell::new(resolved_type_arguments),
-        }
-    }
-}
-
-pub trait UnionOrIntersectionTypeInterface: TypeInterface {
-    fn types(&self) -> &[Rc<Type>];
-}
-
-#[derive(Clone, Debug)]
-#[type_type(interfaces = "UnionOrIntersectionTypeInterface, ObjectFlagsTypeInterface")]
-pub enum UnionOrIntersectionType {
-    UnionType(UnionType),
-}
-
-#[derive(Clone, Debug)]
-#[type_type(impl_from = false)]
-pub struct BaseUnionOrIntersectionType {
-    _type: BaseType,
-    pub types: Vec<Rc<Type>>,
-    pub object_flags: Cell<ObjectFlags>,
-}
-
-impl BaseUnionOrIntersectionType {
-    pub fn new(base_type: BaseType, types: Vec<Rc<Type>>, object_flags: ObjectFlags) -> Self {
-        Self {
-            _type: base_type,
-            types,
-            object_flags: Cell::new(object_flags),
-        }
-    }
-}
-
-impl UnionOrIntersectionTypeInterface for BaseUnionOrIntersectionType {
-    fn types(&self) -> &[Rc<Type>] {
-        &self.types
-    }
-}
-
-impl ObjectFlagsTypeInterface for BaseUnionOrIntersectionType {
-    fn object_flags(&self) -> ObjectFlags {
-        self.object_flags.get()
-    }
-
-    fn set_object_flags(&self, object_flags: ObjectFlags) {
-        self.object_flags.set(object_flags);
-    }
-}
-
-#[derive(Clone, Debug)]
-#[type_type(
-    ancestors = "UnionOrIntersectionType",
-    interfaces = "UnionOrIntersectionTypeInterface, ObjectFlagsTypeInterface"
-)]
-pub struct UnionType {
-    _union_or_intersection_type: BaseUnionOrIntersectionType,
-}
-
-impl UnionType {
-    pub fn new(union_or_intersection_type: BaseUnionOrIntersectionType) -> Self {
-        Self {
-            _union_or_intersection_type: union_or_intersection_type,
-        }
-    }
-}
-
-pub trait ResolvedTypeInterface {
-    fn members(&self) -> Rc<RefCell<SymbolTable>>;
-    fn properties(&self) -> RefMut<Vec<Rc<Symbol>>>;
-    fn set_properties(&self, properties: Vec<Rc<Symbol>>);
-}
-
-#[derive(Clone, Debug)]
-#[type_type]
-pub struct TypeParameter {
-    _type: BaseType,
-    pub constraint: RefCell<Option<Weak<Type>>>, // TODO: is it correct that this is weak?
-    pub is_this_type: Option<bool>,
-}
-
-impl TypeParameter {
-    pub fn new(base_type: BaseType) -> Self {
-        Self {
-            _type: base_type,
-            constraint: RefCell::new(None),
-            is_this_type: None,
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum TypeMapper {
-    Simple(TypeMapperSimple),
-    Array(TypeMapperArray),
-    Function(TypeMapperFunction),
-    Composite(TypeMapperCompositeOrMerged),
-    Merged(TypeMapperCompositeOrMerged),
-}
-
-#[derive(Clone, Debug)]
-pub struct TypeMapperSimple {
-    pub source: Rc<Type>, // TODO: should all of these be weak?
-    pub target: Rc<Type>,
-}
-
-#[derive(Clone, Debug)]
-pub struct TypeMapperArray {
-    pub sources: Vec<Rc<Type>>,
-    pub targets: Option<Vec<Rc<Type>>>,
-}
-
-#[derive(Clone)]
-pub struct TypeMapperFunction {
-    pub func: fn(&TypeChecker, &Type) -> Rc<Type>,
-}
-
-impl fmt::Debug for TypeMapperFunction {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str("TypeMapperFunction")
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct TypeMapperCompositeOrMerged {
-    pub mapper1: Box<TypeMapper>,
-    pub mapper2: Box<TypeMapper>,
-}
-
-impl TypeMapper {
-    pub fn new_simple(source: Rc<Type>, target: Rc<Type>) -> Self {
-        Self::Simple(TypeMapperSimple { source, target })
-    }
-
-    pub fn new_array(sources: Vec<Rc<Type>>, targets: Option<Vec<Rc<Type>>>) -> Self {
-        Self::Array(TypeMapperArray { sources, targets })
-    }
-
-    pub fn new_function(func: fn(&TypeChecker, &Type) -> Rc<Type>) -> Self {
-        Self::Function(TypeMapperFunction { func })
-    }
-
-    pub fn new_composite(mapper1: TypeMapper, mapper2: TypeMapper) -> Self {
-        Self::Composite(TypeMapperCompositeOrMerged {
-            mapper1: Box::new(mapper1),
-            mapper2: Box::new(mapper2),
-        })
-    }
-
-    pub fn new_merged(mapper1: TypeMapper, mapper2: TypeMapper) -> Self {
-        Self::Merged(TypeMapperCompositeOrMerged {
-            mapper1: Box::new(mapper1),
-            mapper2: Box::new(mapper2),
-        })
-    }
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum Ternary {
-    False = 0,
-    Unknown = 1,
-    Maybe = 3,
-    True = -1,
-}
-
-impl TryFrom<i32> for Ternary {
-    type Error = ();
-
-    fn try_from(value: i32) -> Result<Self, Self::Error> {
-        match value {
-            value if value == Ternary::False as i32 => Ok(Ternary::False),
-            value if value == Ternary::Unknown as i32 => Ok(Ternary::Unknown),
-            value if value == Ternary::Maybe as i32 => Ok(Ternary::Maybe),
-            value if value == Ternary::True as i32 => Ok(Ternary::True),
-            _ => Err(()),
-        }
-    }
-}
-
-impl BitAndAssign for Ternary {
-    fn bitand_assign(&mut self, rhs: Self) {
-        *self = (*self as i32 & rhs as i32).try_into().unwrap();
-    }
-}
-
-#[derive(Debug)]
-pub struct ParsedCommandLine {
-    pub file_names: Vec<String>,
-}
-
-pub struct CreateProgramOptions<'config> {
-    pub root_names: &'config [String],
-}
-
-#[non_exhaustive]
-pub struct CharacterCodes;
-#[allow(non_upper_case_globals)]
-impl CharacterCodes {
-    pub const max_ascii_character: char = '';
-
-    pub const line_feed: char = '\n';
-    pub const carriage_return: char = '\r';
-    pub const line_separator: char = ' ';
-    pub const paragraph_separator: char = ' ';
-
-    pub const space: char = ' ';
-
-    pub const underscore: char = '_';
-    pub const dollar_sign: char = '$';
-
-    pub const _0: char = '0';
-    pub const _1: char = '1';
-    pub const _2: char = '2';
-    pub const _3: char = '3';
-    pub const _4: char = '4';
-    pub const _5: char = '5';
-    pub const _6: char = '6';
-    pub const _7: char = '7';
-    pub const _8: char = '8';
-    pub const _9: char = '9';
-
-    pub const a: char = 'a';
-    pub const b: char = 'b';
-    pub const c: char = 'c';
-    pub const d: char = 'd';
-    pub const e: char = 'e';
-    pub const f: char = 'f';
-    pub const g: char = 'g';
-    pub const h: char = 'h';
-    pub const i: char = 'i';
-    pub const j: char = 'j';
-    pub const k: char = 'k';
-    pub const l: char = 'l';
-    pub const m: char = 'm';
-    pub const n: char = 'n';
-    pub const o: char = 'o';
-    pub const p: char = 'p';
-    pub const q: char = 'q';
-    pub const r: char = 'r';
-    pub const s: char = 's';
-    pub const t: char = 't';
-    pub const u: char = 'u';
-    pub const v: char = 'v';
-    pub const w: char = 'w';
-    pub const x: char = 'x';
-    pub const y: char = 'y';
-    pub const z: char = 'z';
-
-    pub const A: char = 'A';
-    pub const B: char = 'B';
-    pub const C: char = 'C';
-    pub const D: char = 'D';
-    pub const E: char = 'E';
-    pub const F: char = 'F';
-    pub const G: char = 'G';
-    pub const H: char = 'H';
-    pub const I: char = 'I';
-    pub const J: char = 'J';
-    pub const K: char = 'K';
-    pub const L: char = 'L';
-    pub const M: char = 'M';
-    pub const N: char = 'N';
-    pub const O: char = 'O';
-    pub const P: char = 'P';
-    pub const Q: char = 'Q';
-    pub const R: char = 'R';
-    pub const S: char = 'S';
-    pub const T: char = 'T';
-    pub const U: char = 'U';
-    pub const V: char = 'V';
-    pub const W: char = 'W';
-    pub const X: char = 'X';
-    pub const Y: char = 'Y';
-    pub const Z: char = 'Z';
-
-    pub const asterisk: char = '*';
-    pub const at: char = '@';
-    pub const backslash: char = '\\';
-    pub const backtick: char = '`';
-    pub const bar: char = '|';
-    pub const close_brace: char = '}';
-    pub const close_bracket: char = ']';
-    pub const close_paren: char = ')';
-    pub const colon: char = ':';
-    pub const comma: char = ',';
-    pub const double_quote: char = '"';
-    pub const equals: char = '=';
-    pub const greater_than: char = '>';
-    pub const hash: char = '#';
-    pub const less_than: char = '<';
-    pub const open_brace: char = '{';
-    pub const open_bracket: char = '[';
-    pub const open_paren: char = '(';
-    pub const plus: char = '+';
-    pub const semicolon: char = ';';
-    pub const single_quote: char = '\'';
-    pub const slash: char = '/';
-}
-
-pub trait ModuleResolutionHost {
-    fn read_file(&self, file_name: &str) -> Option<String>;
-}
-
-pub trait CompilerHost: ModuleResolutionHost {
-    fn get_source_file(&self, file_name: &str) -> Option<Rc<SourceFile>>;
-    fn get_current_directory(&self) -> String;
-    fn get_canonical_file_name(&self, file_name: &str) -> String;
-}
-
-bitflags! {
-    pub struct EmitFlags: u32 {
-        const None = 0;
-        const NoAsciiEscaping = 1 << 24;
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct DiagnosticMessage {
-    pub key: &'static str,
-    pub category: DiagnosticCategory,
-    pub code: u32,
-    pub message: &'static str,
-}
-
-#[derive(Clone, Debug)]
-pub struct DiagnosticMessageChain {
-    pub message_text: String,
-    pub code: u32,
-    pub next: Option<Vec<DiagnosticMessageChain>>,
-}
-
-impl DiagnosticMessageChain {
-    pub fn new(message_text: String, code: u32, next: Option<Vec<DiagnosticMessageChain>>) -> Self {
-        Self {
-            message_text,
-            code,
-            next,
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum Diagnostic {
-    DiagnosticWithLocation(DiagnosticWithLocation),
-    DiagnosticWithDetachedLocation(DiagnosticWithDetachedLocation),
-}
-
-impl Diagnostic {
-    pub fn as_diagnostic_with_detached_location(&self) -> &DiagnosticWithDetachedLocation {
-        enum_unwrapped!(self, [Diagnostic, DiagnosticWithDetachedLocation])
-    }
-}
-
-pub trait DiagnosticInterface: DiagnosticRelatedInformationInterface {
-    fn maybe_related_information(&self) -> Option<&[Rc<DiagnosticRelatedInformation>]>;
-    fn set_related_information(
-        &mut self,
-        related_information: Vec<Rc<DiagnosticRelatedInformation>>,
-    );
-}
-
-#[derive(Clone, Debug)]
-pub struct BaseDiagnostic {
-    _diagnostic_related_information: BaseDiagnosticRelatedInformation,
-    related_information: Option<Vec<Rc<DiagnosticRelatedInformation>>>,
-}
-
-impl BaseDiagnostic {
-    pub fn new(
-        diagnostic_related_information: BaseDiagnosticRelatedInformation,
-        related_information: Option<Vec<Rc<DiagnosticRelatedInformation>>>,
-    ) -> Self {
-        Self {
-            _diagnostic_related_information: diagnostic_related_information,
-            related_information,
-        }
-    }
-}
-
-impl DiagnosticRelatedInformationInterface for BaseDiagnostic {
-    fn maybe_as_diagnostic(&self) -> Option<&Diagnostic> {
-        panic!("Shouldn't call maybe_as_diagnostic() on BaseDiagnostic")
-    }
-
-    fn code(&self) -> u32 {
-        self._diagnostic_related_information.code()
-    }
-
-    fn file(&self) -> Option<Rc<SourceFile>> {
-        self._diagnostic_related_information.file()
-    }
-
-    fn start(&self) -> isize {
-        self._diagnostic_related_information.start()
-    }
-
-    fn length(&self) -> isize {
-        self._diagnostic_related_information.length()
-    }
-
-    fn message_text(&self) -> &DiagnosticMessageText {
-        self._diagnostic_related_information.message_text()
-    }
-}
-
-impl DiagnosticInterface for BaseDiagnostic {
-    fn maybe_related_information(&self) -> Option<&[Rc<DiagnosticRelatedInformation>]> {
-        self.related_information.as_deref()
-    }
-
-    fn set_related_information(
-        &mut self,
-        related_information: Vec<Rc<DiagnosticRelatedInformation>>,
-    ) {
-        self.related_information = Some(related_information);
-    }
-}
-
-impl DiagnosticRelatedInformationInterface for Diagnostic {
-    fn maybe_as_diagnostic(&self) -> Option<&Diagnostic> {
-        Some(self)
-    }
-
-    fn code(&self) -> u32 {
-        match self {
-            Diagnostic::DiagnosticWithLocation(diagnostic_with_location) => {
-                diagnostic_with_location.code()
-            }
-            Diagnostic::DiagnosticWithDetachedLocation(diagnostic_with_detached_location) => {
-                diagnostic_with_detached_location.code()
-            }
-        }
-    }
-
-    fn file(&self) -> Option<Rc<SourceFile>> {
-        match self {
-            Diagnostic::DiagnosticWithLocation(diagnostic_with_location) => {
-                diagnostic_with_location.file()
-            }
-            Diagnostic::DiagnosticWithDetachedLocation(diagnostic_with_detached_location) => {
-                diagnostic_with_detached_location.file()
-            }
-        }
-    }
-
-    fn start(&self) -> isize {
-        match self {
-            Diagnostic::DiagnosticWithLocation(diagnostic_with_location) => {
-                diagnostic_with_location.start()
-            }
-            Diagnostic::DiagnosticWithDetachedLocation(diagnostic_with_detached_location) => {
-                diagnostic_with_detached_location.start()
-            }
-        }
-    }
-
-    fn length(&self) -> isize {
-        match self {
-            Diagnostic::DiagnosticWithLocation(diagnostic_with_location) => {
-                diagnostic_with_location.length()
-            }
-            Diagnostic::DiagnosticWithDetachedLocation(diagnostic_with_detached_location) => {
-                diagnostic_with_detached_location.length()
-            }
-        }
-    }
-
-    fn message_text(&self) -> &DiagnosticMessageText {
-        match self {
-            Diagnostic::DiagnosticWithLocation(diagnostic_with_location) => {
-                diagnostic_with_location.message_text()
-            }
-            Diagnostic::DiagnosticWithDetachedLocation(diagnostic_with_detached_location) => {
-                diagnostic_with_detached_location.message_text()
-            }
-        }
-    }
-}
-
-impl DiagnosticInterface for Diagnostic {
-    fn maybe_related_information(&self) -> Option<&[Rc<DiagnosticRelatedInformation>]> {
-        match self {
-            Diagnostic::DiagnosticWithLocation(diagnostic_with_location) => {
-                diagnostic_with_location.maybe_related_information()
-            }
-            Diagnostic::DiagnosticWithDetachedLocation(diagnostic_with_detached_location) => {
-                diagnostic_with_detached_location.maybe_related_information()
-            }
-        }
-    }
-
-    fn set_related_information(
-        &mut self,
-        related_information: Vec<Rc<DiagnosticRelatedInformation>>,
-    ) {
-        match self {
-            Diagnostic::DiagnosticWithLocation(diagnostic_with_location) => {
-                diagnostic_with_location.set_related_information(related_information)
-            }
-            Diagnostic::DiagnosticWithDetachedLocation(diagnostic_with_detached_location) => {
-                diagnostic_with_detached_location.set_related_information(related_information)
-            }
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum DiagnosticMessageText {
-    String(String),
-    DiagnosticMessageChain(DiagnosticMessageChain),
-}
-
-impl From<String> for DiagnosticMessageText {
-    fn from(string: String) -> Self {
-        DiagnosticMessageText::String(string)
-    }
-}
-
-impl From<DiagnosticMessageChain> for DiagnosticMessageText {
-    fn from(diagnostic_message_chain: DiagnosticMessageChain) -> Self {
-        DiagnosticMessageText::DiagnosticMessageChain(diagnostic_message_chain)
-    }
-}
-
-pub trait DiagnosticRelatedInformationInterface {
-    fn maybe_as_diagnostic(&self) -> Option<&Diagnostic>;
-    fn code(&self) -> u32;
-    fn file(&self) -> Option<Rc<SourceFile>>;
-    fn start(&self) -> isize;
-    fn length(&self) -> isize;
-    fn message_text(&self) -> &DiagnosticMessageText;
-}
-
-#[derive(Clone, Debug)]
-pub enum DiagnosticRelatedInformation {
-    BaseDiagnosticRelatedInformation(BaseDiagnosticRelatedInformation),
-    Diagnostic(Diagnostic),
-}
-
-impl DiagnosticRelatedInformation {
-    pub fn as_diagnostic_with_detached_location(&self) -> &DiagnosticWithDetachedLocation {
-        enum_unwrapped!(
-            self,
-            [
-                DiagnosticRelatedInformation,
-                Diagnostic,
-                DiagnosticWithDetachedLocation
-            ]
-        )
-    }
-}
-
-impl DiagnosticRelatedInformationInterface for DiagnosticRelatedInformation {
-    fn maybe_as_diagnostic(&self) -> Option<&Diagnostic> {
-        match self {
-            DiagnosticRelatedInformation::BaseDiagnosticRelatedInformation(
-                base_diagnostic_related_information,
-            ) => base_diagnostic_related_information.maybe_as_diagnostic(),
-            DiagnosticRelatedInformation::Diagnostic(diagnostic) => {
-                diagnostic.maybe_as_diagnostic()
-            }
-        }
-    }
-
-    fn code(&self) -> u32 {
-        match self {
-            DiagnosticRelatedInformation::BaseDiagnosticRelatedInformation(
-                base_diagnostic_related_information,
-            ) => base_diagnostic_related_information.code(),
-            DiagnosticRelatedInformation::Diagnostic(diagnostic) => diagnostic.code(),
-        }
-    }
-
-    fn file(&self) -> Option<Rc<SourceFile>> {
-        match self {
-            DiagnosticRelatedInformation::BaseDiagnosticRelatedInformation(
-                base_diagnostic_related_information,
-            ) => base_diagnostic_related_information.file(),
-            DiagnosticRelatedInformation::Diagnostic(diagnostic) => diagnostic.file(),
-        }
-    }
-
-    fn start(&self) -> isize {
-        match self {
-            DiagnosticRelatedInformation::BaseDiagnosticRelatedInformation(
-                base_diagnostic_related_information,
-            ) => base_diagnostic_related_information.start(),
-            DiagnosticRelatedInformation::Diagnostic(diagnostic) => diagnostic.start(),
-        }
-    }
-
-    fn length(&self) -> isize {
-        match self {
-            DiagnosticRelatedInformation::BaseDiagnosticRelatedInformation(
-                base_diagnostic_related_information,
-            ) => base_diagnostic_related_information.length(),
-            DiagnosticRelatedInformation::Diagnostic(diagnostic) => diagnostic.length(),
-        }
-    }
-
-    fn message_text(&self) -> &DiagnosticMessageText {
-        match self {
-            DiagnosticRelatedInformation::BaseDiagnosticRelatedInformation(
-                base_diagnostic_related_information,
-            ) => base_diagnostic_related_information.message_text(),
-            DiagnosticRelatedInformation::Diagnostic(diagnostic) => diagnostic.message_text(),
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct BaseDiagnosticRelatedInformation {
-    code: u32,
-    file: Option<Weak<SourceFile>>,
-    start: isize,
-    length: isize,
-    message_text: DiagnosticMessageText,
-}
-
-impl BaseDiagnosticRelatedInformation {
-    pub fn new<TDiagnosticMessageText: Into<DiagnosticMessageText>>(
-        code: u32,
-        file: Option<Rc<SourceFile>>,
-        start: isize,
-        length: isize,
-        message_text: TDiagnosticMessageText,
-    ) -> Self {
-        Self {
-            code,
-            file: file.map(|file| Rc::downgrade(&file)),
-            start,
-            length,
-            message_text: message_text.into(),
-        }
-    }
-}
-
-impl DiagnosticRelatedInformationInterface for BaseDiagnosticRelatedInformation {
-    fn maybe_as_diagnostic(&self) -> Option<&Diagnostic> {
-        None
-    }
-
-    fn code(&self) -> u32 {
-        self.code
-    }
-
-    fn file(&self) -> Option<Rc<SourceFile>> {
-        self.file.as_ref().map(|weak| weak.upgrade().unwrap())
-    }
-
-    fn start(&self) -> isize {
-        self.start
-    }
-
-    fn length(&self) -> isize {
-        self.length
-    }
-
-    fn message_text(&self) -> &DiagnosticMessageText {
-        &self.message_text
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct DiagnosticWithLocation {
-    _diagnostic: BaseDiagnostic,
-}
-
-impl DiagnosticWithLocation {
-    pub fn new(base_diagnostic: BaseDiagnostic) -> Self {
-        Self {
-            _diagnostic: base_diagnostic,
-        }
-    }
-
-    pub fn file_unwrapped(&self) -> Rc<SourceFile> {
-        self.file().unwrap()
-    }
-}
-
-impl DiagnosticRelatedInformationInterface for DiagnosticWithLocation {
-    fn maybe_as_diagnostic(&self) -> Option<&Diagnostic> {
-        panic!("Shouldn't call maybe_as_diagnostic() on DiagnosticWithLocation")
-    }
-
-    fn code(&self) -> u32 {
-        self._diagnostic.code()
-    }
-
-    fn file(&self) -> Option<Rc<SourceFile>> {
-        self._diagnostic.file()
-    }
-
-    fn start(&self) -> isize {
-        self._diagnostic.start()
-    }
-
-    fn length(&self) -> isize {
-        self._diagnostic.length()
-    }
-
-    fn message_text(&self) -> &DiagnosticMessageText {
-        self._diagnostic.message_text()
-    }
-}
-
-impl DiagnosticInterface for DiagnosticWithLocation {
-    fn maybe_related_information(&self) -> Option<&[Rc<DiagnosticRelatedInformation>]> {
-        self._diagnostic.maybe_related_information()
-    }
-
-    fn set_related_information(
-        &mut self,
-        related_information: Vec<Rc<DiagnosticRelatedInformation>>,
-    ) {
-        self._diagnostic
-            .set_related_information(related_information)
-    }
-}
-
-impl From<DiagnosticWithLocation> for Diagnostic {
-    fn from(diagnostic_with_location: DiagnosticWithLocation) -> Self {
-        Diagnostic::DiagnosticWithLocation(diagnostic_with_location)
-    }
-}
-
-impl From<DiagnosticWithLocation> for DiagnosticRelatedInformation {
-    fn from(diagnostic_with_location: DiagnosticWithLocation) -> Self {
-        DiagnosticRelatedInformation::Diagnostic(Diagnostic::DiagnosticWithLocation(
-            diagnostic_with_location,
-        ))
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct DiagnosticWithDetachedLocation {
-    _diagnostic: BaseDiagnostic,
-    pub file_name: String,
-}
-
-impl DiagnosticWithDetachedLocation {
-    pub fn new(base_diagnostic: BaseDiagnostic, file_name: String) -> Self {
-        Self {
-            _diagnostic: base_diagnostic,
-            file_name,
-        }
-    }
-}
-
-impl DiagnosticRelatedInformationInterface for DiagnosticWithDetachedLocation {
-    fn maybe_as_diagnostic(&self) -> Option<&Diagnostic> {
-        panic!("Shouldn't call maybe_as_diagnostic() on DiagnosticWithDetachedLocation")
-    }
-
-    fn code(&self) -> u32 {
-        self._diagnostic.code()
-    }
-
-    fn file(&self) -> Option<Rc<SourceFile>> {
-        self._diagnostic.file()
-    }
-
-    fn start(&self) -> isize {
-        self._diagnostic.start()
-    }
-
-    fn length(&self) -> isize {
-        self._diagnostic.length()
-    }
-
-    fn message_text(&self) -> &DiagnosticMessageText {
-        self._diagnostic.message_text()
-    }
-}
-
-impl DiagnosticInterface for DiagnosticWithDetachedLocation {
-    fn maybe_related_information(&self) -> Option<&[Rc<DiagnosticRelatedInformation>]> {
-        self._diagnostic.maybe_related_information()
-    }
-
-    fn set_related_information(
-        &mut self,
-        related_information: Vec<Rc<DiagnosticRelatedInformation>>,
-    ) {
-        self._diagnostic
-            .set_related_information(related_information)
-    }
-}
-
-impl From<DiagnosticWithDetachedLocation> for Diagnostic {
-    fn from(diagnostic_with_detached_location: DiagnosticWithDetachedLocation) -> Self {
-        Diagnostic::DiagnosticWithDetachedLocation(diagnostic_with_detached_location)
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-pub enum DiagnosticCategory {
-    Warning,
-    Error,
-    Suggestion,
-    Message,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum EmitHint {
-    Expression,
-    Unspecified,
-}
-
-pub struct NodeFactory {}
