+++ src/compiler/types/lines_6000_8000.rs
@@ -1,0 +1,156 @@
+#![allow(non_upper_case_globals)]
+
+use bitflags::bitflags;
+use std::cell::{Cell, Ref, RefCell, RefMut};
+use std::collections::HashMap;
+use std::convert::{TryFrom, TryInto};
+use std::fmt;
+use std::ops::BitAndAssign;
+use std::ops::Deref;
+use std::rc::{Rc, Weak};
+
+use super::{
+    DiagnosticCollection, ModuleSpecifierResolutionHost, Node, NodeId, PseudoBigInt,
+    RelationComparisonResult, SourceFile, SymbolTracker,
+};
+use crate::{NodeBuilder, Number, SortedArray, WeakSelf};
+use local_macros::{ast_type, enum_unwrapped, symbol_type, type_type};
+
+#[derive(Debug)]
+pub struct ParsedCommandLine {
+    pub file_names: Vec<String>,
+}
+
+pub struct CreateProgramOptions<'config> {
+    pub root_names: &'config [String],
+}
+
+#[non_exhaustive]
+pub struct CharacterCodes;
+#[allow(non_upper_case_globals)]
+impl CharacterCodes {
+    pub const max_ascii_character: char = '';
+
+    pub const line_feed: char = '\n';
+    pub const carriage_return: char = '\r';
+    pub const line_separator: char = ' ';
+    pub const paragraph_separator: char = ' ';
+
+    pub const space: char = ' ';
+
+    pub const underscore: char = '_';
+    pub const dollar_sign: char = '$';
+
+    pub const _0: char = '0';
+    pub const _1: char = '1';
+    pub const _2: char = '2';
+    pub const _3: char = '3';
+    pub const _4: char = '4';
+    pub const _5: char = '5';
+    pub const _6: char = '6';
+    pub const _7: char = '7';
+    pub const _8: char = '8';
+    pub const _9: char = '9';
+
+    pub const a: char = 'a';
+    pub const b: char = 'b';
+    pub const c: char = 'c';
+    pub const d: char = 'd';
+    pub const e: char = 'e';
+    pub const f: char = 'f';
+    pub const g: char = 'g';
+    pub const h: char = 'h';
+    pub const i: char = 'i';
+    pub const j: char = 'j';
+    pub const k: char = 'k';
+    pub const l: char = 'l';
+    pub const m: char = 'm';
+    pub const n: char = 'n';
+    pub const o: char = 'o';
+    pub const p: char = 'p';
+    pub const q: char = 'q';
+    pub const r: char = 'r';
+    pub const s: char = 's';
+    pub const t: char = 't';
+    pub const u: char = 'u';
+    pub const v: char = 'v';
+    pub const w: char = 'w';
+    pub const x: char = 'x';
+    pub const y: char = 'y';
+    pub const z: char = 'z';
+
+    pub const A: char = 'A';
+    pub const B: char = 'B';
+    pub const C: char = 'C';
+    pub const D: char = 'D';
+    pub const E: char = 'E';
+    pub const F: char = 'F';
+    pub const G: char = 'G';
+    pub const H: char = 'H';
+    pub const I: char = 'I';
+    pub const J: char = 'J';
+    pub const K: char = 'K';
+    pub const L: char = 'L';
+    pub const M: char = 'M';
+    pub const N: char = 'N';
+    pub const O: char = 'O';
+    pub const P: char = 'P';
+    pub const Q: char = 'Q';
+    pub const R: char = 'R';
+    pub const S: char = 'S';
+    pub const T: char = 'T';
+    pub const U: char = 'U';
+    pub const V: char = 'V';
+    pub const W: char = 'W';
+    pub const X: char = 'X';
+    pub const Y: char = 'Y';
+    pub const Z: char = 'Z';
+
+    pub const asterisk: char = '*';
+    pub const at: char = '@';
+    pub const backslash: char = '\\';
+    pub const backtick: char = '`';
+    pub const bar: char = '|';
+    pub const close_brace: char = '}';
+    pub const close_bracket: char = ']';
+    pub const close_paren: char = ')';
+    pub const colon: char = ':';
+    pub const comma: char = ',';
+    pub const double_quote: char = '"';
+    pub const equals: char = '=';
+    pub const greater_than: char = '>';
+    pub const hash: char = '#';
+    pub const less_than: char = '<';
+    pub const open_brace: char = '{';
+    pub const open_bracket: char = '[';
+    pub const open_paren: char = '(';
+    pub const plus: char = '+';
+    pub const semicolon: char = ';';
+    pub const single_quote: char = '\'';
+    pub const slash: char = '/';
+}
+
+pub trait ModuleResolutionHost {
+    fn read_file(&self, file_name: &str) -> Option<String>;
+}
+
+pub trait CompilerHost: ModuleResolutionHost {
+    fn get_source_file(&self, file_name: &str) -> Option<Rc<SourceFile>>;
+    fn get_current_directory(&self) -> String;
+    fn get_canonical_file_name(&self, file_name: &str) -> String;
+}
+
+bitflags! {
+    pub struct EmitFlags: u32 {
+        const None = 0;
+        const NoAsciiEscaping = 1 << 24;
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum EmitHint {
+    Expression,
+    Unspecified,
+}
+
+pub struct NodeFactory {}
