+++ src/compiler/types/mod.rs
@@ -1,6 +1,18 @@
-mod lines_0_4000;
-pub use lines_0_4000::*;
-mod lines_4000_8000;
-pub use lines_4000_8000::*;
+mod lines_0_1000;
+pub use lines_0_1000::*;
+mod lines_1000_2000;
+pub use lines_1000_2000::*;
+mod lines_2000_4000;
+pub use lines_2000_4000::*;
+mod lines_4000_5000;
+pub use lines_4000_5000::*;
+mod lines_5000_5250;
+pub use lines_5000_5250::*;
+mod lines_5250_5500;
+pub use lines_5250_5500::*;
+mod lines_5500_6000;
+pub use lines_5500_6000::*;
+mod lines_6000_8000;
+pub use lines_6000_8000::*;
 mod lines_8000_end;
 pub use lines_8000_end::*;
